src/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
bcrypt==4.0.0
billiard==3.6.4.0
blinker==1.5
Brotli==1.0.9
cached-property==1.5.2
celery==5.1.2
cffi==1.15.1
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.4.1
dataclasses==0.8
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
Flask-WeasyPrint==0.6
Flask-WTF==1.0.1
fonttools==4.27.1
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.1.0
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
Pillow==8.4.0
prompt-toolkit==3.0.31
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.9
pyphen==0.11.0
pytz==2022.2.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.41
tinycss2==1.1.1
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.3
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.1
WTForms==3.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.1.9
